File and Bundle refactoring
