+++ rollup.config.js
@@ -14,15 +14,17 @@ const globals = {
 };
 
 export default [
-  bundle('src/index.js',         'dist/angular-vue.js',         'AngularVue'),
-  bundle('src/plugins/index.js', 'dist/angular-vue-plugins.js', 'AngularVuePlugins'),
-
-  bundle('src/directives/vue.js',                    'dist/directives/ng-vue.js',            'AngularVueDirective'),
-  bundle('src/components/angular-vue.js',            'dist/components/ng-vue.js',            'AngularVueComponent'),
+  bundle('src/index.js',                             'dist/index.js',                        'AngularVue'),
+  bundle('src/directives/ng-vue.js',                 'dist/directives/ng-vue.js',            'AngularVueDirective'),
+  bundle('src/components/ng-vue.js',                 'dist/components/ng-vue.js',            'AngularVueComponent'),
   bundle('src/plugins/angular-vue-plugin.js',        'dist/plugins/ng-vue-plugin.js',        'AngularVuePlugin'),
   bundle('src/plugins/angular-vue-plain-plugin.js',  'dist/plugins/ng-vue-plain-plugin.js',  'AngularVuePlainPlugin'),
   bundle('src/plugins/angular-vue-route-plugin.js',  'dist/plugins/ng-vue-route-plugin.js',  'AngularVueRoutePlugin'),
   bundle('src/plugins/angular-vue-router-plugin.js', 'dist/plugins/ng-vue-router-plugin.js', 'AngularVueRouterPlugin'),
+
+  //Legacy Compatibility
+  bundle('src/ng-vue-module.js',  'dist/angular-vue.js',         'AngularVueModule'),
+  bundle('src/plugins/index.js',  'dist/angular-vue-plugins.js', 'AngularVuePlugins'),
 ];
 
 function bundle(inFile, outFile, name) {
