+++ src/ng-vue-module.js
@@ -1,0 +1,6 @@
+import angular from 'angular';
+import vueDirective from './directives/ng-vue.js';
+
+const ngModule = angular.module('angularVue', []);
+
+vueDirective.register(ngModule);
